pyproject.toml
[tool.poetry]
name = "puma"
version = "0.1.0"
description = ""
authors = ["guimassoqueto <gui.massoqueto@gmail.com>"]
readme = "README.md"
packages = [
{include = "*.py"},
{include = "app/**/*.py"}
]

[tool.poetry.dependencies]
python = "^3.10"
selectolax = "^0.3.16"
playwright = "^1.37.0"
aiohttp = "^3.8.5"
pydantic = "^2.3.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.10"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
